Accordion svg fix
